moved smartRouteLogic to its own file
